+++ packages/expo-maps/CHANGELOG.md
@@ -4,13 +4,17 @@
 
 ### 🛠 Breaking changes
 
-- [Android] Add zoom, tilt, bearing to StreetView position ([#35938](https://github.com/expo/expo/pull/35938) by [@jakex7](https://github.com/jakex7))
-
 ### 🎉 New features
 
 ### 🐛 Bug fixes
 
 ### 💡 Others
+
+## 0.9.0 — 2025-04-08
+
+### 🛠 Breaking changes
+
+- [Android] Add zoom, tilt, bearing to StreetView position ([#35938](https://github.com/expo/expo/pull/35938) by [@jakex7](https://github.com/jakex7))
 
 ## 0.8.0 — 2025-04-04